Update strings. Package rename
